FilterExps

DataAssimilationBenchmarks.FilterExps.filter_param — Method

filter_param(args::Tuple{String,String,Int64,Int64,Float64,Int64,Float64,Float64, 
                         Float64,Int64,Float64,Float64})

Filter joint state-parameter estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments.

DataAssimilationBenchmarks.FilterExps.filter_state — Method

filter_state(args::Tuple{String,String,Int64,Int64,Float64,Int64,Float64, 
                         Int64,Float64})

Filter state estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments.