Global Types

DataAssimilationBenchmarks.ArView — Type

ArView = Union{Array{Float64, 2}, SubArray{Float64, 2}}

Type union of arrays and views of arrays for use within ensemble conditioning operations, integration schemes and other array operations.

DataAssimilationBenchmarks.ConM — Type

ConM = Union{UniformScaling{Float64}, Symmetric{Float64}}

Type union of conditioning matrix types, which will be used for optimization routines in the transform method.

DataAssimilationBenchmarks.CovM — Type

CovM = Union{UniformScaling{Float64}, Diagonal{Float64}, Symmetric{Float64}}

Type union of covariance matrix types, for optimized computation based on characteristics.

DataAssimilationBenchmarks.ParamDict — Type

ParamDict = Union{Dict{String, Array{Float64}}, Dict{String, Vector{Float64}}}

Dictionary for model parameters to be passed to derivative functions by name.

DataAssimilationBenchmarks.TransM — Type

TransM = Union{Tuple{Symmetric{Float64,Array{Float64,2}},Array{Float64,2},Array{Float64,2}},
               Tuple{Symmetric{Float64,Array{Float64,2}},Array{Float64,1},Array{Float64,2}},
               Array{Float64,2}}

Type union of right transform types, including soley a transform, or a transform, weights and rotation package.

DataAssimilationBenchmarks.VecA — Type

VecA = Union{Vector{Float64}, SubArray{Float64, 1}}

Type union of vectors and ensemble members of sample for using in integration schemes and related array operations.

DataAssimilationBenchmarks.ParamSample — Type

ParamSample = Dict{String, Vector{UnitRange{Int64}}}

Dictionary containing key and index pairs to subset the state vector and then merge with dx_params in parameter estimation problems.