Differential Equation Solvers

DataAssimilationBenchmarks.DeSolvers.ParamDict — Type

ParamDict = Union{Dict{String, Array{Float64}}, Dict{String, Vector{Float64}}}

Dictionary for model parameters to be passed to derivative functions by name.

DataAssimilationBenchmarks.DeSolvers.VecA — Type

VecA = Union{Vector{Float64}, SubArray{Float64, 1}}

Type union of vectors and ensemble members of sample for using in integration schemes and related array operations.

DataAssimilationBenchmarks.DeSolvers.ParamSample — Type

ParamSample = Dict{String, Vector{UnitRange{Int64}}}

Dictionary containing key and index pairs to subset the state vector and then merge with dx_params in parameter estimation problems.

DataAssimilationBenchmarks.DeSolvers.em_step! — Method

em_step!(x::Vector{Float64}, t::Float64, kwargs::Dict{String,Any})

This will propagate the state x one step forward by Euler-Murayama scheme. Step size is h, the Wiener process is assumed to have a scalar diffusion coefficient.

DataAssimilationBenchmarks.DeSolvers.rk4_step! — Method

rk4_step!(x::VecA, t::Float64, kwargs::Dict{String,Any})

Step of integration rule for 4 stage Runge-Kutta as discussed in Grudzien et al. 2020. The rule has strong convergence order 1.0 for generic SDEs and order 4.0 for ODEs. Arguments are given as:

x            -- array or sub-array of a single state possibly including parameter values
t            -- time point
kwargs       -- should include dx_dt, the paramters for the dx_dt and optional arguments
dx_dt        -- time derivative function with arguments x and dx_params
dx_params    -- parameters necessary to resolve dx_dt, not including
                parameters to be estimated in the extended state vector 
h            -- numerical discretization step size
diffusion    -- tunes the standard deviation of the Wiener process, 
                equal to sqrt(h) * diffusion
diff_mat     -- structure matrix for the diffusion coefficients,
                replaces the default uniform scaling 
state_dim    -- keyword for parameter estimation, dimension of the
                dynamic state < dimension of full extended state
param_sample -- ParamSample dictionary for merging extended state with dx_params
ξ            -- random array size state_dim, can be defined in kwargs
                to provide a particular realization for method validation

DataAssimilationBenchmarks.DeSolvers.tay2_step! — Method

tay2_step!(x::Vector{Float64}, t::Float64, kwargs::Dict{String,Any})

Deterministic second order (autonomous) Taylor method for step size h and state vector x. Time variable t is just a dummy variable, where this method is not defined for non-autonomous dynamics. Arguments are given as x – array of a single state possibly including parameter values kwargs – should include dxdt, the paramters for the dxdt and optional arguments dxdt – time derivative function with arguments x and dxparams dxparams – parameters necessary to resolve dxdt, not including parameters to be estimated in the extended state vector h – numerical discretization step size