Lorenz-96 model

DataAssimilationBenchmarks.L96.dx_dt — Method

dx_dt(x::VecA, t::Float64, dx_params::ParamDict)

Time derivative for Lorenz-96 model, x is a single model state of size state_dim, t
is a dummy time argument for consistency with integration methods, dx_params is a 
parameter dictionary which is called for the forcing parameter.

DataAssimilationBenchmarks.L96.jacobian — Method

jacobian(x::Vector{Float64}, t::Float64, dx_params::ParamDict) 

Computes the Jacobian of Lorenz-96 about the state x. The time variable t is a dummy
variable for consistency with integration methods, dx_params is a parameter dictionary
which is called for the forcing parameter. Note that this is designed to load entries in
a zeros array and return a sparse array to make a compromise between memory
and computational resources.

DataAssimilationBenchmarks.L96.l96s_tay2_step! — Method

l96s_tay2_step!(x::Vector{Float64}, t::Float64, kwargs::Dict{String,Any}) 

One step of integration rule for l96 second order taylor rule
The ρ and α are to be computed by the auxiliary functions in the L96 submodule, depending
only on p, and supplied for all steps. This is the general formulation which includes,
eg. dependence on the truncation of terms in the auxilliary function C with
respect to the parameter p.  In general, truncation at p=1 is all that is
necessary for order 2.0 convergence, and in this case C below is identically
equal to zero.  This auxilliary function can be removed (and is removed) in other
implementations for simplicity.

This method is derived in
Grudzien, C. et al.: On the numerical integration of the Lorenz-96 model,
with scalar additive noise, for benchmark twin experiments,
Geosci. Model Dev., 13, 1903–1924, https://doi.org/10.5194/gmd-13-1903-2020, 2020.
NOTE: this Julia version still pending validation as in the above manuscript