D3VARExps

This module defines methods for experiments with classical variational data assimilation with 3D-VAR. Primal cost functions are defined, with their implicit differentiation performed with automatic differentiation with JuliaDiff methods. Development of gradient-based optimization schemes using automatic differentiation is ongoing, with future development planned to integrate variational benchmark experiments.

The basic 3D-VAR cost function API is defined as follows

    D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T),
    obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real

where the control variable x is optimized, with fixed hyper-parameters defined in a wrapping function passed to auto-differentiation.

Methods

DataAssimilationBenchmarks.D3VARExps.D3_var_filter_analysis — Method

function D3varfilteranalysis((timeseries, γ, isinformed, tuningfactor, isupdated)::NamedTuple{ (:timeseries,:γ,:isinformed,:tuningfactor,:is_updated),<:Tuple{String, Float64,Bool,Float64,Bool}}) Plotting capabilities are commented out for parallel experiment.

DataAssimilationBenchmarks.D3VARExps.D3_var_filter_analysis_simple — Method

D3_var_filter_analysis_simple()