FilterExps

FilterExps API

The FilterExps.jl sub-module configures twin experiments using stored time series data as generated above for efficiency when using the same base-line time series to generate possibly different experiment configurations. Experiment configurations are generated by

filter_state(args::Tuple{String,String,Int64,Int64,Float64,Int64,Float64,Int64,Float64})
time_series, method, seed, nanl, obs_un, obs_dim, γ, N_ens, infl = args

where time_series specifies the path to the .jld2 truth twin, method specifies the filter scheme, seed specifies the pseudo-random seed, nanl is the number of observation / analysis times to produce a posterior estimate for, obs_un specifies the observation error standard deviation, assuming a uniform scaling observation error covariance, obs_dim specifies the dimension of the observation vector, γ specifies the level of the nonliearity in the alternating_obs_operator, N_ens specifies the ensemble size and infl specifies the (static) multiplicative inflation.

Similar conventions follow for parameter estimation experiments

filter_param(args::Tuple{String,String,Int64,Int64,Float64,Int64,
                         Float64,Float64,Float64,Int64,Float64,Float64})
time_series, method, seed, nanl, obs_un, obs_dim,
γ, param_err, param_wlk, N_ens, state_infl, param_infl = args

with the exception of including the standard deviation, param_err, of the initial iid Gaussian draw of the parameter sample centered at the true value, the standard deviation, param_wlk, of the random walk applied to the parameter sample after each analysis and the multiplicative covariance inflation applied separately to the extended parameter states alone param_infl.

Docstrings

DataAssimilationBenchmarks.FilterExps.filter_param — Method

filter_param((time_series::String, method::String, seed::Int64, nanl::Int64, 
              obs_un::Float64, obs_dim::Int64, γ::Float64, p_err::Float64, p_wlk::Float64,
              N_ens::Int64, s_infl::Float64, p_infl::Float64)::NamedTuple)

Filter joint state-parameter estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments of the NamedTuple.

DataAssimilationBenchmarks.FilterExps.filter_state — Method

filter_state((time_series::String, method::String, seed::Int64, nanl::Int64,
              obs_un::Float64, obs_dim::Int64, γ::Float64, N_ens::Int64,
              s_infl::Float64)::NamedTuple)

Filter state estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments of the NamedTuple.