SingleExperimentDriver

Following the convention of GenerateTimeSeries, FilterExps and SmootherExps using NamedTuple arguments to define hyper-parameter configurations, the SingleExperimentDriver module defines dictionaries of the form

experiment_group["parameter_settings"]

where keyword arguments return standard parameter configurations for these experiments with known results for reproducibility. These standard configurations are used in the package for for debugging, testing, benchmarking and profiling code. Package tests use these standard configurations to verify a DA method's forecast and analysis RMSE. User-defined, custom experiments can be modeled from the methods in the above modules with a corresponding SingleExperimentDriver dictionary entry used to run and debug the experiment, and to test and document the expected results. Parallel submission scripts are used for production runs of sensitivity experiments, defined in ParallelExperimentDriver.

Experiment groups

DataAssimilationBenchmarks.SingleExperimentDriver.classic_enks_exps — Constant

classic_enks_exps = Dict{String, Any}

Pre-configured classic EnKS style twin experiments for benchmark models.

DataAssimilationBenchmarks.SingleExperimentDriver.d3_var_exps — Constant

d3_var_exps = Dict{String, Any}

Pre-configured 3D-VAR style twin experiments for benchmark models.

DataAssimilationBenchmarks.SingleExperimentDriver.enkf_exps — Constant

enkf_exps = Dict{String, Any}

Pre-configured EnKF style twin experiments for benchmark models.

DataAssimilationBenchmarks.SingleExperimentDriver.ienks_exps — Constant

ienks_exps = Dict{String, Any}

Pre-configured IEnKS style twin experiments for benchmark models.

DataAssimilationBenchmarks.SingleExperimentDriver.sienks_exps — Constant

sienks_exps = Dict{String, Any}

Pre-configured SIEnKS style twin experiments for benchmark models.

DataAssimilationBenchmarks.SingleExperimentDriver.time_series_exps — Constant

time_series_exps = Dict{String, Any}

Pre-configured time series experiments for generating test data for twin experiments.