SmootherExps

SmootherExps API

The SmootherExps.jl sub-module configures twin experiments using stored time series data as generated above for efficiency when using the same base-line time series to generate possibly different experiment configurations. Experiment configurations are generated by function calls as with the filter experiments, but with the additional options of how the outer-loop is configured with a classic, single-iteration or the fully iterative Gauss-Newton style smoother. The parameters lag and shift specify how the data assimilation windows are translated in over the observation and analysis times. The mda parameter is only applicable to the single-iteration and Gauss-Newton style smoothers, utlizing sequential multiple data assimilation. Note, the single-iteration and fully iterative Gauss-Newton style smoothers are only defined for MDA compatible values of lag and shift where the lag is an integer multiple of the shift.

Currently debugged and validated smoother experiment configurations include

classic_state          -- classic EnKS style state estimation
classic_param          -- classic EnKS style state-parameter estimation
single_iteration_state -- single-iteration EnKS state estimation
single_iteration_param -- single-iteration EnKS state-parameter estimation
iterative_state        -- Gauss-Newton style state estimation
iterative_param        -- Gauss-Newton style state-parameter estimation

Other techniques are still in debugging and validation. Each of these takes an analysis type as used in the transform function in the EnsembleKalmanSchemes.jl sub-module, like the filter analyses in the filter experiments.

All experiments are funcitonalized so that they can be called from an array of parameter values which will typically be varied with naive parallelism. Relevant arguments and experimental results are dumped as a side effect to a dictionary in a JLD2. Experiments return their runtime in minutes.

Docstrings

DataAssimilationBenchmarks.SmootherExps.classic_param — Method

classic_param((time_series::String, method::String, seed::Int64, nanl::Int64, lag::Int64,
               shift::Int64, obs_un::Float64, obs_dim::Int64, γ::Float64, p_err::Float64,
               p_wlk::Float64, N_ens::Int64, s_infl::Float64,
               s_infl::Float64})::NamedTuple)

Classic ensemble Kalman smoother joint state-parameter estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments. NOTE: the classic scheme does not use multiple data assimilation and we hard code mda=false in the function for consistency with other methods.

DataAssimilationBenchmarks.SmootherExps.classic_state — Method

classic_state((time_series::String, method::String, seed::Int64, nanl::Int64, lag::Int64,
               shift::Int64, obs_un::Float64, obs_dim::Int64, γ::Float64, N_ens::Int64, 
               s_infl::Float64)::NamedTuple)

Classic ensemble Kalman smoother state estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments. NOTE: the classic scheme does not use multiple data assimilation and we hard code mda=false in the function for consistency with other methods.

DataAssimilationBenchmarks.SmootherExps.iterative_param — Method

iterative_param((time_series:String, method:String, seed::Int64, nanl::Int64, lag::Int64,
                 shift::Int64, mda::Bool, obs_un::Float64, obs_dim::Int64, γ::Float64,
                 p_err::Float64, p_wlk::Float64, N_ens::Int64,
                 s_infl::Float64, p_infl::Float64})

4DEnVAR joint state-parameter estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments.

DataAssimilationBenchmarks.SmootherExps.iterative_state — Method

iterative_state(args::Tuple{String,String,Int64,Int64,Int64,Int64,Bool,Float64, 
                            Int64,Float64,Int64,Float64})

4DEnVAR state estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments.

DataAssimilationBenchmarks.SmootherExps.single_iteration_param — Method

single_iteration_param((time_series::String, method::String, seed:Int64, nanl::Int64,
                        lag::Int64, shift::Int64, mda::Bool, obs_un::Float64,
                        obs_dim::Int64, γ::Float64, p_err::Float64, p_wlk::Float64,
                        N_ens::Int64, s_infl::Float64, p_infl::Float64)::NamedTuple)

SIEnKS joint state-parameter estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments.

DataAssimilationBenchmarks.SmootherExps.single_iteration_state — Method

single_iteration_state((time_series::String, method::String, seed::Int64, nanl::Int64,
                        lag::Int64, shift::Int64, mda::Bool, obs_un::Float64,
                        obs_dim::Int64, γ::Float64, N_ens::Int64,
                        s_infl::Float64})::NamedTuple)

SIEnKS state estimation twin experiment. Twin experiment parameters such as the observation dimension, observation uncertainty, data assimilation method, number of cycles, ensemble size etc. are specified in the arguments.