Ensemble Kalman Schemes

API for data assimilation solvers

Different filter and smoothing schemes are run through the routines including

ensemble_filter(analysis::String, ens::Array{Float64,2}, obs::Vector{Float64},
                         obs_cov::CovM, state_infl::Float64, kwargs::Dict{String,Any})

ls_smoother_classic(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},
                             obs_cov::CovM, state_infl::Float64, kwargs::Dict{String,Any})

ls_smoother_single_iteration(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},
                             obs_cov::CovM, state_infl::Float64, kwargs::Dict{String,Any})

ls_smoother_gauss_newton(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},
                             obs_cov::CovM, state_infl::Float64, kwargs::Dict{String,Any};
                             ϵ::Float64=0.0001, tol::Float64=0.001, max_iter::Int64=10)

# type union for multiple dispatch over specific types of covariance matrices
CovM = Union{UniformScaling{Float64}, Diagonal{Float64}, Symmetric{Float64}}

analysis   -- string of the DA scheme string name, given to the transform sub-routine in
              methods
ens        -- ensemble matrix defined by the array with columns given by the replicates of
              the model state
obs        -- observation vector for the current analysis in ensemble_filter / array with
              columns given by the observation vectors for the ordered sequence of analysis
							times in the current smoothing window
obs_cov    -- observation error covariance matrix, must be positive definite of type CovM
state_infl -- multiplicative covariance inflation factor for the state variable covariance
              matrix, set to one this is the standard Kalman-like update
kwargs     -- keyword arguments for parameter estimation or other functionality, including
              integration parameters for the state model in smoothing schemes

The type of analysis to be passed to the transform step is specified with the analysis string. Observations for the filter schemes correspond to information available at a single analysis time while the ls (lag-shift) smoothers require an array of observations corresponding to all analysis times within the DAW. Observation covariances should be typed according to the type union above for efficiency. The state_infl is a required tuneable parameter for multiplicative covariance inflation. Extended parameter state covariance inflation can be specified in kwargs. These outer loops will pass the required values to the transform function that generates the ensemble transform for conditioning on observations. Different outer-loop schemes can be built around the transform function alone in order to use validated ensemble transform schemes. Utility scripts to generate observation operators, analyze ensemble statistics, etc, are included in the EnsembleKalmanSchemes.jl sub-module. See the experiments directory discussed below for example usage.

Docstrings

DataAssimilationBenchmarks.EnsembleKalmanSchemes.alternating_obs_operator — Method

alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64, 
                         kwargs::Dict{String,Any})

Observation of alternating state vector components, possibly nonlinear transformation. This selects components to observe based on the observation dimension and if parameter estimation is being performed. Parameters are always unobservable, and even states will be removed from the state vector until the observation dimension is appropriate. Nonlinear observations are optional, as described for the Lorenz-96 model by Asch, Bocquet, Nodet pg. 181.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.analyze_ens — Method

analyze_ens(ens::ArView, truth::Vector{Float64})

Computes the ensemble RMSE as compared with truth twin, and the ensemble spread.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.analyze_ens_para — Method

analyze_ens_para(ens::ArView, truth::Vector{Float64})

Computes the ensemble RMSE as compared with truth twin, and the ensemble spread.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.ens_update! — Method

ens_update!(ens::ArView, transform::T1) where {T1 <: TransM}

Updates ensemble by right-transform method. In the case where this follows the stochastic EnKF as in Carrassi et al. 2018, this simply performs right mutliplication. All other cases use the 3-tuple including the right transform for the anomalies, the weights for the mean and the random, mean-preserving orthogonal matrix.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.ensemble_filter — Method

ensemble_filter(analysis::String, ens::Array{Float64,2}, obs::Vector{Float64},  
                obs_cov::CovM, s_infl::Float64, kwargs::Dict{String,Any})

General filter analysis step, wrapping the transform / update, and inflation steps. Optional keyword argument includes state_dim for extended state including parameters. In this case, a value for the parameter covariance inflation should be included in addition to the state covariance inflation.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.inflate_param! — Method

inflate_param!(ens::Array{Float64,2}, inflation::Float64, sys_dim::Int64, 
               state_dim::Int64)

Parameter variable multiplicative covariance inflation. State variables are assumed to be in the leading rows, while extended state, parameter variables are after. Multiplicative inflation is performed only in the trailing components.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.inflate_state! — Method

inflate_state!(ens::Array{Float64,2}, inflation::Float64, sys_dim::Int64, 
               state_dim::Int64)

Dynamic state variable multiplicative covariance inflation. State variables are assumed to be in the leading rows, while extended state variables, parameter variables are after. Multiplicative inflation is performed only in the leading components.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.ls_smoother_classic — Method

ls_smoother_classic(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},  
                    obs_cov::CovM, s_infl::Float64, kwargs::Dict{String,Any})

Lag-shift ensemble kalman smoother analysis step, classical version. Classic enks uses the last filtered state for the forecast, different from the iterative schemes which use the once or multiple-times re-analized posterior for the initial condition for the forecast of the states to the next shift.

Optional argument includes state dimension for extended state including parameters. In this case, a value for the parameter covariance inflation should be included in addition to the state covariance inflation.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.ls_smoother_gauss_newton — Method

ls_smoother_gauss_newton(analysis::String, ens::Array{Float64,2},                   
                         obs::Array{Float64,2}, obs_cov::CovM, s_infl::Float64,
                         kwargs::Dict{String,Any}; ϵ::Float64=0.0001,
                         tol::Float64=0.001, max_iter::Int64=5)

Lag-shift Gauss-Newton IEnKS analysis step, algorithm 4, Bocquet & Sakov 2014 ienks uses the final re-analyzed posterior initial state for the forecast, which is pushed forward in time from the initial conidtion to shift-number of observation times. Optional argument includes state dimension for an extended state including parameters. In this case, a value for the parameter covariance inflation should be included in addition to the state covariance inflation.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.ls_smoother_single_iteration — Method

ls_smoother_single_iteration(analysis::String, ens::Array{Float64,2},       
                             obs::Array{Float64,2}, obs_cov::CovM,
                             s_infl::Float64, kwargs::Dict{String,Any})

Lag-shift ensemble kalman smoother analysis step, single iteration version. Single-iteration enks uses the final re-analyzed posterior initial state for the forecast, which is pushed forward in time to shift-number of observation times. Optional argument includes state dimension for an extended state including parameters. In this case, a value for the parameter covariance inflation should be included in addition to the state covariance inflation.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.rand_orth — Method

rand_orth(N_ens::Int64)

This generates a random. mean preserving, orthogonal matrix as in Sakov & Oke 2008.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.square_root — Method

square_root(M::T) where {T <: CovM}

Auxiliary function for computing the square roots of multiple types of covariance matrices, with the subroutines defined according to the sub-type.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.square_root_inv — Method

square_root_inv(M::T) where {T <: CovM}

Auxiliary function for computing the square root inverse of multiple types of covariance matrices, with the subroutines defined according to the sub-type. This will optionally return a computation of the inverse and the square root itself all as a byproduct of the singular value decomposition computation.

DataAssimilationBenchmarks.EnsembleKalmanSchemes.transform — Method

transform(analysis::String, ens::Array{Float64,2}, obs::Vector{Float64},        
       obs_cov::CovM, kwargs::Dict{String,Any}; conditioning::ConM=1000.0I, 
       m_err::Array{Float64,2}=(1.0 ./ zeros(1,1)),
       tol::Float64 = 0.0001,
       j_max::Int64=40,
       Q::CovM=1.0I)

Computes transform and related values for various flavors of ensemble Kalman schemes. Serves as an auxilliary function for EnKF, ETKF(-N), EnKS, ETKS(-N), IEnKS(-N), where "analysis" is a string which determines the type of transform update. The observation error covariance should be of UniformScaling, Diagonal or Symmetric type.