IEEE39 Bus Swing Equation Model

DataAssimilationBenchmarks.IEEE39bus.dx_dt — Method

dx_dt(x::VecA, t::Float64, dx_params::ParamDict)

Time derivative of the phase and fequency of the effective-network swing equation model. Input x is a 2 n_g VecA of the phase and fequency at each of the n_g generator buses. The input dx_params of type ParamDict containing system parameters to be passed to the integration scheme. The system is currenty defined autonomously to be run as an SDE, noise perturbed steady state.