Global Types

The following types are declared for optimizing numerical routines, for using multiple dispatch of functions with different specified input forms and for passing arguments between inner / outer loop steps of the DA twin experiment.

DataAssimilationBenchmarks.ParamSample — Type

ParamSample = Dict{String, Vector{UnitRange{Int64}}}

Dictionary containing key and index pairs to subset the state vector and then merge a statistical sample of parameters that govern the equations of motion with the ParamDict dx_params in parameter estimation problems.

DataAssimilationBenchmarks.StepKwargs — Type

StepKwargs = Dict{String,Any}

Key word arguments for twin experiment time stepping. Arguments are given as:

    REQUIRED:
    dx_dt        -- time derivative function with arguments x and dx_params
    dx_params    -- parameters necessary to resolve dx_dt, not including
                    parameters to be estimated in the extended state vector 
    h            -- numerical time discretization step size
    γ            -- controls nonlinearity of the alternating_obs_operator

    OPTIONAL:
    diffusion    -- tunes the standard deviation of the Wiener process, 
                    equal to sqrt(h) * diffusion
    diff_mat     -- structure matrix for the diffusion coefficients,
                    replaces the default uniform scaling 
    state_dim    -- keyword for parameter estimation, dimension of the
                    dynamic state < dimension of full extended state
    param_sample -- ParamSample dictionary for merging extended state with dx_params
    ξ            -- random array size state_dim, can be defined in kwargs
                    to provide a particular realization for method validation

See DataAssimilationBenchmarks.EnsembleKalmanSchemes.alternating_obs_operator for a discusssion of the γ parameter.

DataAssimilationBenchmarks.ArView — Method

ArView = Union{Array{T, 2}, SubArray{T, 2}} where T <: Real

Type union of Arrays and SubArrays for use within ensemble conditioning operations, integration schemes and other array operations.

DataAssimilationBenchmarks.ConM — Method

ConM = Union{UniformScaling{T}, Symmetric{T}} where T <: Real

Type union of conditioning matrix types, which are used for optimization routines in the transform method.

DataAssimilationBenchmarks.CovM — Method

CovM = Union{UniformScaling{T}, Diagonal{T}, Symmetric{T}} where T <: Real

Type union of covariance matrix types, for optimized computation based on their special characteristics as symmetric, positive definite operators.

DataAssimilationBenchmarks.ParamDict — Method

ParamDict = Union{Dict{String, Array{T}}, Dict{String, Vector{T}}} where T <: Real

Dictionary for model parameters to be passed to derivative functions by name. This allows one to pass both vector parameters (and scalars written as vectors), as well as matrix valued parameters such as diffusion arrays.

DataAssimilationBenchmarks.VecA — Method

VecA = Union{Vector{T}, SubArray{T, 1}} where T <: Real

Type union of Vectors and 1-D SubArrays in order to pass columns of an ensemble maxtrix into integration schemes and related array operations.