GenerateTimeSeries

GenerateTimeSeries is a sub-module used to generate a time series for a twin experiment based on tuneable model configuration parameters. Example syntax for the configuration of a time series is as follows, where arguments are defined in a named tuple to be passed to the specific experiment function:

    (seed::Int64, h::Float64, state_dim::Int64, tanl::Float64, nanl::Int64, spin::Int64,
     diffusion::Float64)::NamedTuple)

The seed argument specifies initial condition for the pseudo-random number generator on which various simulation settings will depend – simulations will be reproduceable with the same seed value. The numerical integration step size is given by the argument h which controls the discretization error of the numerically simulated experiment. The size of the Lorenz-96 system scales with the state_dim argument, though other models such as the IEEE39bus model are of pre-defined size. The length of continuous time units between sequential observations is controlled with the tanl (time-between-analysis) argument which defines the frequency of pseudo data outputs. The number of observations / analyses to be saved is controlled with the nanl (number-of-analyses) argument. The length of the spin-up for the integration of the dynamical system solution to guarantee a stationary observation generating process is controlled with the spin argument. The diffusion parameter determining the intensity of the random perturbations is controlled with the diffusion argument. Results are saved in .jld2 format in the data directory to be called by filter / smoother experiments cycling over the pseudo-observations.

Docstrings

DataAssimilationBenchmarks.GenerateTimeSeries.IEEE39bus_time_series — Method

IEEE39bus_time_series((seed::Int64, h:Float64, tanl::Float64, nanl::Int64, spin::Int64,
                       diffusion::Float64)::NamedTuple)

Simulate a "free run" time series of the IEEE39 Bus Swing Equation Model for generating an observation process and truth twin for data assimilation twin experiments. Output from the experiment is saved in a dictionary of the form,

Dict{String, Any}(
                  "seed" => seed,
                  "h" => h,
                  "diffusion" => diffusion,
                  "diff_mat" => diff_mat,
                  "dx_params" => dx_params, 
                  "tanl" => tanl,
                  "nanl" => nanl,
                  "spin" => spin,
                  "obs" => obs,
                  "model" => "IEEE39bus"
                 )

Where the file name is written dynamically according to the selected parameters as follows:

"IEEE39bus_time_series_seed_" * lpad(seed, 4, "0") * 
"_diff_" * rpad(diffusion, 5, "0") * 
"_tanl_" * rpad(tanl, 4, "0") * 
"_nanl_" * lpad(nanl, 5, "0") * 
"_spin_" * lpad(spin, 4, "0") * 
"_h_" * rpad(h, 5, "0") * 
".jld2"

DataAssimilationBenchmarks.GenerateTimeSeries.L96_time_series — Method

L96_time_series((seed::Int64, h::Float64, state_dim::Int64, tanl::Float64, nanl::Int64,
                 spin::Int64, diffusion::Float64, F::Float64)::NamedTuple)

Simulate a "free run" time series of the Lorenz-96 model model for generating an observation process and truth twin for data assimilation twin experiments. Output from the experiment is saved in a dictionary of the form,

Dict{String, Any}(
                  "seed" => seed,
                  "h" => h,
                  "diffusion" => diffusion, 
                  "dx_params" => dx_params, 
                  "tanl" => tanl,
                  "nanl" => nanl,
                  "spin" => spin,
                  "state_dim" => state_dim,
                  "obs" => obs,
                  "model" => "L96"
                 )

Where the file name is written dynamically according to the selected parameters as follows:

"L96_time_series_seed_" * lpad(seed, 4, "0") * 
"_dim_" * lpad(state_dim, 2, "0") * 
"_diff_" * rpad(diffusion, 5, "0") * 
"_F_" * lpad(F, 4, "0") * 
"_tanl_" * rpad(tanl, 4, "0") * 
"_nanl_" * lpad(nanl, 5, "0") * 
"_spin_" * lpad(spin, 4, "0") * 
"_h_" * rpad(h, 5, "0") * 
".jld2"